Added final gameplay elements, improved main menu functionality, added audio files
