--- README.md
@@ -1,1 +1,0 @@
-# space_invaders_final_project
