+++ scoreboard.txt
@@ -1,0 +1,5 @@
+Fer: 55
+Kobe: 56
+Jordan: 63
+Ben: 10
+Bird: 33
